Invert search direction (search up the directory tree, not down)
Note: This change is not backwards-compatible. When switching
to an upward-searching version of cite, all existing bibtex
files need to be moved for cite to keep working.

The previous behaviour was to search subdirectories for bibtex
files. This is the inverse direction from what is commonly used
as a search path for configuration files. E. g. .gitignore and
.editorconfig files are both taken from parent, rather than from
subdirectories. Using parent directories should also make shared
bibtex files easier to use.